chore: mobile navbar wip
